suites options ui logic to renderer process
