+++ src/main-process/updates.ts
@@ -51,7 +51,7 @@ const showNoUpdatesDialog = async (): Promise<void> => {
 
     const noUpdatesDialogOpts: Electron.MessageBoxOptions = {
       buttons: ['OK'],
-      detail: 'You’ve got the latest version of Loadmill Desktop Agent',
+      detail: 'You’ve got the latest version of Loadmill Desktop App',
       message: 'Your app is up to date',
       type: 'info',
     };
