+++ src/types/suite.ts
@@ -2,3 +2,5 @@ export type SuiteOption = {
   description: string;
   id: string;
 };
+
+export type TestSuite = SuiteOption & { [key: string]: unknown };
